requirements.txt
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-ESL==1.4.18
Werkzeug==2.0.1
WTForms==2.3.3
